Switched to an in-memory system.
The library now operates on an in-memory system where changes have to be
manually flushed to the file system using <UltraStarFile>.flush(). This
solves many problems where the file could theoretically be modified
during editing. Writes are as atomic as they can be, however, they
obviously overwrite changed made since parsing the file.